mcplotloader bugfix: always convert xvals, yvals and errvals to float, never allow strings here. Any string might get +'d in js later on as strings rather than as numbers, causing NaN.

+++ tools/Python/mccodelib/mcplotloader.py
@@ -400,13 +400,13 @@ def _load_multiplot_1D_lst(f_dat):
         for l in lines:
             if '#' in l:
                 continue
-            xvals.append(l.split()[0])
+            xvals.append(float(l.split()[0]))
 
             for i in range(len(yvars)//2):
                 yvals_lst.append([])
                 yvals_err_lst.append([])
-                yvals_lst[i].append(l.split()[2*i+num_xvars])
-                yvals_err_lst[i].append(l.split()[2*i+num_xvars+1])
+                yvals_lst[i].append(float(l.split()[2*i+num_xvars]))
+                yvals_err_lst[i].append(float(l.split()[2*i+num_xvars+1]))
         header.xvals = xvals
         
         # create a new instance for each y variable
